// src/shared/redux/models/IArtistDetails.tsx
export interface IArtistDetails {
  albums:  { [key: string]: Album };
  artists: Artists;
  songs:   { [key: string]: Song };
 }
 
 export interface Album {
  attributes?: AlbumAttributes;
  href:        string;
  id:          string;
  type:        Type;
 }
 
 export interface AlbumAttributes {
  albumName?:            string;
  artistName:            Name;
  artwork:               Artwork;
  audioLocale?:          AudioLocale;
  audioTraits:           AudioTrait[];
  composerName?:         ComposerName;
  copyright?:            string;
  discNumber?:           number;
  durationInMillis?:     number;
  editorialNotes?:       PurpleEditorialNotes;
  genreNames:            string[];
  hasLyrics?:            boolean;
  hasTimeSyncedLyrics?:  boolean;
  isAppleDigitalMaster?: boolean;
  isCompilation?:        boolean;
  isComplete?:           boolean;
  isMasteredForItunes:   boolean;
  isPrerelease?:         boolean;
  isSingle?:             boolean;
  isrc?:                 string;
  name:                  string;
  playParams:            PlayParams;
  previews?:             Preview[];
  recordLabel?:          RecordLabel;
  releaseDate:           Date;
  trackCount?:           number;
  trackNumber?:          number;
  upc?:                  string;
  url:                   string;
 }
// ...
 export enum Type {
  Albums = "albums",
  Songs = "songs",
 }
 
 export interface Artists {
  [key: string]: The136975;
 }
 
 export interface The136975 {
  attributes:    The136975_Attributes;
  href:          string;
  id:            string;
  meta:          The136975_Meta;
  relationships: Relationships;
  type:          string;
  views:         The136975_Views;
 }
// ...
 export interface Song {
  attributes?: SongAttributes;
  href:        string;
  id:          string;
  meta?:       SongMeta;
  type:        Type;
 }
 
 export interface SongAttributes {
  albumName?:            string;
  artistName:            Name;
  artwork:               Artwork;
  audioLocale?:          AudioLocale;
  audioTraits:           AudioTrait[];
  composerName?:         ComposerName;
  copyright?:            string;
  discNumber?:           number;
  durationInMillis?:     number;
  editorialNotes?:       FluffyEditorialNotes;
  genreNames:            GenreName[];
  hasLyrics?:            boolean;
  hasTimeSyncedLyrics?:  boolean;
  isAppleDigitalMaster?: boolean;
  isCompilation?:        boolean;
  isComplete?:           boolean;
  isMasteredForItunes:   boolean;
  isPrerelease?:         boolean;
  isSingle?:             boolean;
  isrc?:                 string;
  name:                  string;
  playParams:            PlayParams;
  previews?:             Preview[];
  recordLabel?:          RecordLabel;
  releaseDate:           Date;
  trackCount?:           number;
  trackNumber?:          number;
  upc?:                  string;
  url:                   string;
 }
 
 export interface FluffyEditorialNotes {
  short: string;
 }
 
 export interface SongMeta {
  formerIds: string[];
 }